requirements.txt
Flask
flask-cors
python-dotenv
langchain-openai
mcp-use
faker
sqlalchemy
psycopg2
pytest
pytest-asyncio
